split of msg/blocks to be own object/class
